Make scrollToTop a shared function
